Be a little more conservative in handling "changed" events.
Previously, if we got a "changed" event for a path we weren't
monitoring, we would log a warning and ignore it. Now, we log the
warning, but treat it like a "created" event so we know about the
file/directory going forward. This situation shouldn't occur, but this
new logic will handle it a little better.
(cherry picked from commit 29701f299632ea343d7b58af4d3b7a143ced0078)