Metadata: Better handling of floating cert-authenticated clients (Ticket #1030)
From the ticket:

A floating, cert-authenticated client can be not recognized properly by
hostname if it resolves to an arbitrary name in reverse DNS.

Background: Metadata.resolve_client, called from @exposed Core's
methods, falls back to reverse DNS lookup for client's name, because the
name is not preserved thanks to bailing off early from
Metadata.AuthenticateConnection.

(This issue can be related to #936.)

This patch enables caching of client names for cert-based floating
clients.

Signed-off-by: Sol Jerome <sol.jerome@gmail.com>