bcfg2: Ignore "-l" if configuring from a file
The "-f" option, which tells the client to configure from a file rather
than querying the server, cannot be combined with the "-l" option (nor
with the equivalent setting in the bcfg2.conf(5) file), as a decision
list won't be available.  Instead of crashing (with a KeyError:
'decision_list'), the client will now ignore the "-l" option if the "-f"
option also is specified.