Improved exit codes for the bcfg2 client
If you're using bcfg2 in a shell script or some other automated way,
it would be nice to know *why* you are getting a non-zero exit code.

This change adds a Bcfg2/Errors.py with a list of exit code values
with descriptive names.

There are some generic exit codes starting at zero and a handful of
client-based exit codes starting at 101.